+++ Changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - The Set types derive `Hash`, `Ord` and `PartialOrd`
 - `to_repr` method to get the internal representation value.
+- Implement `as_bitset` to the base enum. Allows creating a set from a single element easier in `const` context (since `.into()` is not available there).
 
 
 ## [v0.1.1] enum-bitset-derive - 2025-06-08